requirements.txt
fastapi
uvicorn
pandas
requests
openai
python-dotenv
google-analytics-data
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2